backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.2
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
